jarvid_api (Module5-Lesson1)/requirements.txt
fastapi
uvicorn
pydantic
openai
faiss-cpu
numpy
prometheus_client